Add files via upload
This a fully responsive web design which was assigned to me as an assignment for the hackathon held for SMIT batch-11. I created this website using html and css and it is based on 4 sections also this is my first ever fully responsive web design. I hope it finds you in peace and you like it.